src/test/javascript/spec/app/entities/sample/sample-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import SampleUpdateComponent from '@/entities/sample/sample-update.vue';
import SampleClass from '@/entities/sample/sample-update.component';
import SampleService from '@/entities/sample/sample.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Sample Management Update Component', () => {
    let wrapper: Wrapper<SampleClass>;
    let comp: SampleClass;
    let sampleServiceStub: SinonStubbedInstance<SampleService>;

    beforeEach(() => {
      sampleServiceStub = sinon.createStubInstance<SampleService>(SampleService);

      wrapper = shallowMount<SampleClass>(SampleUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          sampleService: () => sampleServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.sample = entity;
        sampleServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(sampleServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.sample = entity;
        sampleServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(sampleServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundSample = { id: 123 };
        sampleServiceStub.find.resolves(foundSample);
        sampleServiceStub.retrieve.resolves([foundSample]);

        // WHEN
        comp.beforeRouteEnter({ params: { sampleId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.sample).toBe(foundSample);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
